build fixes for clang12/libstdc++11
Fix stricter warnings and include issues introduced by this toolchain
combination.

+++ tests/unit/framework/test_init_factory.cpp
@@ -27,5 +27,5 @@ TEST_CASE("verify init factory behavior", "[init_factory]")
     std::vector<std::unique_ptr<test_base>> test_objects;
     int counter = 0;
     test_base::make_all(test_objects, counter);
-    REQUIRE(test_objects.size() == counter);
+    REQUIRE(static_cast<int>(test_objects.size()) == counter);
 }
